Move the API token length in the config file and adjust db accordingly
Up to now the API token length was hard-coded in the code to 30 chars
long. With this commit this limit is now set in the configuration
file of the application.
In order for this to make sense, we need to bump the size of the field
in the database to 255 (should be more than enough).

NOTE: the alembic upgrade won't work on sqlite which consider pretty
much only two types of field: numeric or text. So it has no problem
storing a 100 chars long string in a varchar(40).