Create the API namespace
This is the namespace that we will use to describe and expose the copr API.
This commit introduce it, move some code from misc to the api_general and
start with two API functions to list someone's copr and add a new one.

These two functions are documented in the /api/ (ie in the api.html template).