support for kerberos authentication
The krb auth method is aimed to be self-standing auth method, but
also can be used as alternative way how to connect to Copr
in parallel with Fedora OpenID authentication method.

The OpenID is currently still the default and the Copr 'username'
is bound to OpenID username (such usernames must match);  this
should be fixed in future so that the Copr user is 1:N mapped to
OpenID credentials (similarly like the krb is).