Add optional fedmenu resources to every page.
This is for https://fedoraproject.org/wiki/Fedora_Engineering/FY16_Plan

If the FEDMENU_URL and FEDMENU_DATA_URL options are left out of the
config, then nothing should change on the displayed page.  If they are
included, then they will be pulled in at page load time to render little
menu buttons on the bottom right hand corner of the page.

There's an example up at http://threebean.org/fedmenu/

It will be a while until we are able to turn this on globally across all
our apps.