[backend] abstraction above [BE <-> JG <-> Builders] channels
Previously, if backend started a bit faster than job-grabber, it
was fine, but the other way around (because job-grabber is
prerequisite for backend in service file):

  1. Job Grabber started
  2. JobGrabber took the list of tasks
  3. JobGrabber filled the redis task queue && and filled its
     internal representation of queue
  4. Backend started, which resulted in redis task queue cleanup
  5. JobGrabber was confused because the filled redis queue did
     not match its internal queue representation state

This resulted in unprocessed queue.

Now backend precisely tells job_grabber via "different" (Queue)
channel about backend re-start (so backend does not have to touch
the Queue itself and the inconsistency does not happen).