Vagrantfile: fix the first start of frontend
The Builds and Monitor views tracebacked, as the SQL function
were not created, because we created the new database from
models instead of alembic.

This fix still creates the DB from models, but then downgrades
before the first alembic revision containing SQL functions, and
then upgrades again. This way, we get the functions to the
database.