[copr-dist-git] Don't assume the SCM repo has the same name as the package
Mock(-scm) expects the cloned directory to has the same name as the package
it's building. To make sure it is so we explicitly pass the package name to
git clone as a destination directory name.

This allows (amont other things) to have a single git repo for multiple spec
files.