check for duplicities during creating
without this we do not check for duplicities in group projects when creating over API call.
This regression was introduced in commit 5e102adb07e26b8738166b994a6fc5d74b8ae589.

The strange-looking condition
   models.Copr.id != copr.id
needs to be added there because when we add group to copr model (where group is backref) then
sqlalchemy call imlicit flush() so the new Copr object is sudenly visible.
It is nowhere documented, but I find some other people reporting the same behaviour on internet.