fix for python-progress API
The 'max' is unfortunate name but it needs to be used; otherwise
the graphical part of progress bar does not correctly show the
remaining part of file to be uploaded.

The usage of 'max' argument can be seen in 'class Progress' in
python-progress project.

Revert "W: 36,23: Redefining built-in 'max' (redefined-builtin)"

This reverts commit 42c0942b759d58b22c2c2e5551eb67088cf5d9ac.