[frontend] specify how to create module table, not generate it from module
I dont think that generating is a good idea anymore. It makes versioning
of the table dificult. In case of some modifications of the table,
this script would allways create up-to-date table schema. Which might
be problem because there will also be a migrations for changing
the schema and it will be in conflict. Rather just specify collumns,
that should be created and how.