[backend] drop useless check for existing workers
This seems to be buggy.  Frontend never returns job which is
already assigned to some worker, unless there's some bug..

Removing this check allows us to actually properly react on SSH
connection failure (either sshd died, or whole VM is unusable); so
that Worker (a) terminates the VM and (b) asks Frontend to
resubmit the job again.