[builder] More live logs with sed instead of 'col -b'
There's --lines option for 'col', but it defines the "minimal"
number of lines buffered, not the "maximal" number of lines.  And
in our case, it is enough to keep only one line in buffer.

Previously, the output was printed in chunks.  Now the mock output
should be much more fluent.

Complements: 6b8e426d73cccefd7d0cf3fd5df21429f7170136