[dist-git] Bug 1457888 - Mock SCM method fails to build a package
The newly rewritten method worked a bit differently in the way that
it placed only a repo subdir where spec was placed into the tarball.
In other words, it expected the spec to be next to sources, the same
way Tito expects it. This is not the case for some coprs in production
(e.g. jscotka/modularity-testing-framework)