[python] always send name of the user
This data field for specifying username serves for needs of proxyuser.
I.e when the user owning the project/module/build is different to the user,
who authentifies himself agains the frontend API.

However, it always needs to be an user. If we have e.g. group project,
it is assigned to an user who created it nevertheless.

For handling group projects, we use specific URL and pass group name to it.
The data field is for proxy user purposes and should not affect this.