[backend] msgbus: support for ssl + initial connection failover
The 'host' and 'port' configuration options are obsoleted now in
favor of 'hosts' array (reflects the stomppy implementation).  To
be dropped.

The 'auth' config option now allows user to specify client's
key/crt and cacert bundle.

Slightly modified the connection policy;  worker waits till the
stomp connection is done before it continues with the assigned
task.