[frontend] new generic web-hook
This webhook is theoretically useable for the existing --srpm
build methods, but we need to get the webhook payload parsers (for
Github and Gitlab) concentrated on one place, and get the combo
parser applied here, too (TODO task).

The webhook payload is for now only stored in temporary data
directory, so the end-user (via the future srpm build method) will
have the payload in hand and can parse it manually.