[frontend] custom method: move the script filter into form
This caused problems when (a) the script was specified in "add
package" or "edit package" field (the '\r\n' sequences were kept)
and then re-submit by webhook caused that the broken script was
inherited by build.

Still, don't apply the filter if the first line (shebang) doesn't
contain '\r\n' sequence.  In that case the file is specified
through API or copr-cli and it it means that we don't want to
change even even single byte in the script (and if there are some
'\r\n' sequences later in the file, it could be be desired).

Complements c93546f660ecd37ca61f21bf3ee09fcffc0dacbd