[frontend][python] send json data as a _file_ in FileRequest
We need to do that, because we send files as data via `MultipartEncoder
and it is not possible to combine data= and json= in requests.request.

We could just set a data fields, but then we would lose the advantages
of json (e.g. boolean can't be encoded when MultipartEncoder is used)