[keygen] fix gnupg2 requirement
The code has always used gpg2, which comes from the gnupg2 package on
Fedora and RHEL.  On RHEL, there is only the gnupg2 package and it
provides gnupg as well.  On Fedora, gnupg provides gnupg-1.4.x which
includes /bin/gpg.

By chance, 'Requires: gnupg' does the right thing on RHEL.  On Fedora it
pulls in an unneeded dependency and fails to properly ensure that
/bin/gpg2 is installed.  This is easily overlooked because core tools
like dnf & yum require gnupg2.

Prepare for the coming deprecation of gnupg-1.4.x by making it clear
that gnupg2 is what copr-keygen requires.