[selinux][backend][dist-git][frontend] fix SELinux
All concerned packages now conditionally 'Require: copr-selinux'.
Since RPM 4.14 (at the time of writing this code, it is at least
on Fedora 27+) we can use "boolean deps".  And it exactly matches
our use-case -- we require/install copr-selinux if and only if
selinux-policy-targeted is installed.

So all the 'semanage' ruless from %post scriptlets can go away.

The fixes are also complaint to the latest Independent SELinux
Policy draft (except that the draft suggest us to use %posttrans,
but it's discussed with SELinux guys).

The 'mls' policy is not built for now.

Related: #301