[cli] keeping 'copr mock-config' functionality
No doubts, this functionality is needed and there's no real
technical reason to deprecate this, I already offered a help to
solve this.

There were refused theories that we'll implement this
functionality in 'copr-rpmbuild', but that's not viable option
since copr-rpmbuild is not wanted on EPEL7 (see PR#306), nor even
on EPEL6.

This reverts commit a42861a1150e3467d4347f0f67128b994cad1def.