[prunerepo] fix tests for rawhide
* since dnf3, dnf repoquery --latest-limit=1 returns multiple packages for the
same latest NEVRA. With dnf2, only one of the packages maching the latest NEVRA
was returned.

Practical implication is that prunerepo will not prune any of the packages having
the latest NEVRA so it will keep more things on the disk. This is essentially
a bug. Only the rpm with the latest build time could be kept for each NEVRA.

See https://bugzilla.redhat.com/show_bug.cgi?id=1612469