[frontend] separate manage.py commands to their own files
IMHO the content of manage.py file is getting out of control. I am
moving all the *Command implementations to their own files to gain
better readability and orientation in the code, but also possibility
to implement helper _stuff_ needed for the command.

Some of the commands are quite long and contain some logic.
Such logic needs to be implemented on the class level, which can
be annoying (e.g. some additional helper class needs to be added as
a nested class). Ofc it can be put on the module level and mixed among
command classes, but this solution isn't very good. Another solution
would be putting such code into `coprs.logic` or `coprs.helpers`, but
it doesn't make much sense, if its only purpose is to work within the
command.

Therefore, I am separating the commands into their own files to get
rid of any limitations of their implementation.