[cli] fix side_effect function arguments
When side_effect function is passed, it is called with the same
arguments as the mock function. In this case it is a
`BuildProxy.get(self, build_id)`

However, "same arguments" means something different for current
python and python2.6 which is on epel6. Now if method is mocked,
then side_effect function is called with the method's self as the
first argument. On epel it is not.