[cli] fix the APIv1 call in the aciton_new_webhook_secret function
I don't really know how this happened, but `action_new_webhook_secret`
was implemented in APIv1, using the old client, while the rest of the
copr-cli code was reimplemented to APIv3. That means, that the old
code here thinks that the `self.client` is the old client, that
there is a `parse_name` function and so on.

I am fixing the APIv1 call instead of rewriting it to the APIv3
at once because it is a release time and I have no time to do so.

Though, it should be reimplemented ASAP.