[frontend] stick with the old repo_id format until F28 is supported
We cannot really switch to the new
copr:{hostname}:{owner}:{project} format yet, because it is implemented in
dnf-plugins-core-3.x which is only on F29+

Since the F29+ plugin is able to work with both old and new formats, we can
safely stay with the old one until F28 is still supported. Once it goes EOL,
we can migrate to the new format.