[frontend] use input data without `None` fields
to transform only the specified values to the WTForms, so they
can correctly use the default values for the rest (e.g. when
`spec_template` is not set).

Also, there is some issue when passing an existing package object
(or its data) to the form in order to use it as defaults, so I am
workarounding it.

It turns out that the whole JSON --> WTForms adapter code wasn't
written very well and is more patchy and harder to use,
than it needs to be. We should address that.