[beaker-tests-sanity] do not wait till an action is created
There shouldn't need to be a wait till action is created. Action for gpg key
generation is created in the same transaction with the project itself,
therefore once the `copr-cli create ...` command is finished, there
should be visible pending action.

I am changing this because there is some problem and the `while` loop never
finished and caused the tests to hang out. After this change the tests
fail for me, which is much better. IMHO they fail because backend
finishes the action before tests can even fetch it. I can't confirm
that, but at this time it is not possible to rewrite the test to be
more reliable, so I am ignoring the failure