[beaker-tests-sanity] remove the complicated sed
I must admit that I don't understand the reason why the sed is here.
We wan't to take a part of JSON data (which should be valid JSON by
themselves) from `jq` command, then store it to a file. Then other tests
are getting the data and processing them through `jq` again. I see no
point in modifying them manually with sed.

Moreover, the sed breaks the JSON format and unables `jq` to read the
data. It generates output like this

    pypi_package_name": "motionpaint,

missing the opening quote for key and ending quote for value.