[frontend] return modulemd yaml as a string, not bytes
There are several ways to obtain modulemd yaml before it gets proceeded.
We can request it from an URL, user can upload the file, or we can
generate it from UI form.

Both ``ModuleProvider.from_url`` and ``ModulemdGenerator.generate``
return the yaml content as string, however ``ModuleProvider.from_file``
currently returns it as bytes. This inconsistency makes issues in the
next steps of the pipeline - particularly when reading the yaml
with the third-party class ``Modulemd.ModuleStream.import_from_string``