[python] fix wait function for custom List
The `wait` function is supposed to work with on `waitable` instances.
It should work on a `Munch` returned by other API method (i.e. it has
a `__proxy__` property), or `list` of such munches.

It should also work on a `helpers.List` of `Munch` objects returned
by some API method. In such case, we set the `__proxy__` property
to the `List` not munches themselves.