[frontend] pass queue_sizes also to the graph page
This change fixes /status/stats page, which now fails with:

    jinja2.exceptions.UndefinedError: 'queue_sizes' is undefined.

The issue was caused by PR#590. I am fixing it by adding
a `context_processor` function to the blueprint instead of passing
the variable inside of the `stats()` function. For layout variables
such as badges in the menu, it IMHO makes much more sense
to pass them by blueprint context processor, instead of doing
it in every view.