[frontend] support ?priority=x for non copr:// repo
This is an alternative to PR#539

Initially, I thought, that it will be enough to remove the
copr:// condition from `parse_repo_params` to get the
special params dict and simultaneously keep ?priority=x
in the repo baseurl. However, although baseurl with ?priority
can be opened in the browser, DNF can't handle it, so we need
to strip it.

Theoretically, false positives can occur, but IMHO it is
not a real threat.