[frontend][cli] new --max-builds option
Allow users to configure build garbage collector for packages.
If enabled, the specified number of builds (the newest, by build
id) is kept and the rest is garbage collected (on a daily basis).

Since the garbage collector deletes the oldest copr builds per
build ID, it does different thing from what prunerepo does on it's
own independently (where only the newest RPM (sub) package with
the greates ENVR is kept).

Resolves: #614, PR#620