[frontend] reset config changes after each test method
We use following statements in our tests:

    app.config["foo"] = "bar"

to *temporarily* change and test application behavior.
The problem is, that we don't reset the config changes to their
original values. As a consequence, test methods are not independent
and it matters in what order they are executed, whether some
tests were skipped and issues like #637 appear.

I am chaning this and fixing the tests that rely on this bug.