[frontend] better parse Pagure's PR messages
Pull-requests in Pagure nowadays send also
'*.pull-request.updated' and '*.pull-request.rebased' messages,
so let's use those instead of 'pull-request.comment.added'.

To allow repeated builds per one pull-request, we don't store
pull-request URL to database anymore (we store there a unique
location of object for each push), so it is not easy to reference
the pull request in Pagure UI from Copr UI anymore.  We can fix
this later if anyone considers this to be important.

Fixes: #586, PR#641