[frontend][python] print friendly error on http when https is enforced
See #502

Allow to set boolean field `encrypted` in the copr-cli config.
By default, it is set to `True` and it is not suggested to
a user to explicitly set it in the config. Only on instances,
that don't enforce https (e.g. local docker-compose instance),
it is.

When the connection should be `encrypted` and `copr_url` is
accidentally set to http:// URL, error is raised.