[frontend] make clean_old_builds query join() explicit
The python3-sqlalchemy-1.3.1 on Fedora 30+ is more scrict on
join()s, and it better detected our over-linkage between
CoprDir--Build--Package tables (the broken schema problem is
tracked as tracked as issue #617).

Detected by testsuite on F30:
E sqlalchemy.exc.InvalidRequestError: Can't determine which FROM
clause to join from, there are multiple FROMS which can join to
this entity. Try adding an explicit ON clause to help resolve the
ambiguity.

Fixes: PR#666