[frontend] expect the correct status code for project deletion failure
The c2fcfff for some reason changed the expected status code for a
failed request to remove project (because of unfinished builds in it).
Seems to me like the assert was changed to just comply with a bugged
frontend behavior. Or maybe those two bugs negated each other.

However, once I've fixed the issue with wrong `finished` status for
SRPM builds in ab70329, then this test started failing.