[frontend] fix a failing test
This test IMHO used to pass only because of conjunction of several bugs.
I don't really understand this, but here is my theory.

There wasn't added any chroot for the build in the first place,
so everything that was done for each chroot just trivially passed.
The ab70329 fixed the `finished` status for SRPM builds (which is
basically build without chroots) and therefore this test started
failing. Moreover, I couldn't get working

    [chroot.name for chroot in self.b_few_chroots.build_chroots]

when it was after the frontend request. So I moved it above because
its position doesn't really matter.