[frontend] fix finished status for SRPM builds, hopefully last time
There was a mistake in the test and therefore a finished status
could be wrong without us noticing. The original test didn't really
tested SRPM builds because even though we changed `source_status`,
the build has build_chroots so it wasn't correctly representing
an SRPM build.

I am creating a constant defining the list of finished states,
so it can be easily used within `finished` properties for several
models.