[frontend] notify outdated chroots with 80 days interval
The original 14 days was too short. When having 180 days preservation
period, then it is around 13 notifications for a chroot.
With 80 days interval between notifications, it is 3 notifications
per chroot (one straight away, the second one at day 80, and the
last one at day 160). The remaining 20 days will be highlighted
in red, thanks to the PR#715.