[rpmbuild] builder: merge dependencies from playbooks
The 'copr-builder' package is designed to prepare working Copr
builder machine, so it's better to define all dependencies in the
RPM directly (rather than in Ansible playbooks).

Certain subset of packages is required to be kept on it's latest
version, so when appropriate please use %latest_requires from now.
That macro is used to generate /usr/bin/copr-update-builder
script, which is to be executed anytime Copr build system decides
to do so (especially when a fresh new VM was spawned from image).

The list is taken from:
https://infrastructure.fedoraproject.org/cgit/ansible.git/tree/\
roles/copr/backend/files/provision/provision_builder_tasks.yml\
?id=0704eabc9d6ea31464c442014a68ebe398532350