[backend] add copr_print_results_to_delete.py script
See #712

The PR#719 fixed the issue #712, however chroots that disappeared
because of it are now permanently lost and the mechanism for
deleting outdated chroots will never ever remove their backend data.

I am introducing this script for discovering backend data, that are
not known by frontend (and therefore should be probably removed).
It will not remove the data by itself because there may be
false-positives of what we really want to remove. Therefore, it
just lists the directories.