[docker] run httpd on port 5000 instead of 80
We wouldn't need to do this because of httpd, however there is a
following use-case

    supervisorctl stop httpd
    sudo -u copr-fe python3 /opt/copr/frontend/coprs_frontend/manage.py \
        runserver -p 80 -h 0.0.0.0 --no-reload

which won't work because a non-root user can't start a server on
privileged port (bellow 1024). Therefore, we need to switch to
using 5000 for both `manage.py` server and httpd.

I really dislike this change, so if you know how allow running the
runserver command as `copr-fe` user on port `80`, please do so
and then revert this commit.