[rpmbuild] clean /var/cache/mock automatically
This is to prevent "no space left on device" on / partition.
Since we recycle user-dedicated builder machines heavily, we need
to clean directories like

    $ ls -1 /var/cache/mock
    BUILD_ID-fedora-rawhide-x86_64
    BUILD_ID-fedora-rawhide-x86_64-bootstrap

which contain yum/dnf caches there (soon to be enabled), and
mock's root_cache.

If we don't do this, after several (tens?) of builds on the same
builder, the / partition might be wasted.  This is especially
important in Fedora Copr on aarch64 where we have small disks.