backend: really skip VMs with check failure
This seems to be a bug which could cause weird things like "VM not
responding", but still it was requested to do some task.

Such task could fail without obvious reasons -- and the worst
thing on this would be that Copr would blame the user for the
mistake (failed build).