frontend: nicer submitter identification
.. both in web-UI and in backend's redis database.

For normal builds (submitted via web-UI or API) we still print the
copr user link (pointing to list of user's copr projects).

But for webhooks - where we don't know build.user_id - this isn't
possible.  So if the webhook method allows this (.. so it sets the
'build.submitted_by' field) print that nicely in the web-UI
instead.