backend: don't run appstream-builder for PR dirs
The ':pr:' directories are created dynamically, so we can not
request admins to touch '.disable-appstream' file for every such
dir.  We would have to check '../../<main_dir>' for that file.

But first it would be slightly more complicated, and second it
doesn't make much sense to pay attention to it because the ':pr:'
directories are usually much smaller than the main directory (so
appstream-builder is cheap operation).

This patch is triggered by the same AssertionError as previous
commit.