[frontend] respect the module platform
See #606

Note that this functionality is described differently in documentation than we are
implementing it.

> For example the module can buildrequire platform:[f28, f29] modules, which means
> it should be built against the f28 and f29 streams of platform module.

See "Module Stream Expansion (MSE)" section at
https://pagure.io/fm-orchestrator/blob/master/f/docs/HOW_MBS_BUILDS_MODULES.rst

While we are taking a list of platform streams like ["f28", "f29"] and
submit module builds into the respective chroots that are enabled in a project,
e.g. ["fedora-28-x86_64", "fedora-29-x86_64", "fedora-29-i386"]

This is IMHO the only implementation that makes any sense for Copr,
at least for the time being.