beaker-tests: move to f30 docker environment
We changed the format of repo IDs, and such format only works on
F29+ where is up2date dnf-plugins-core.  See PR#799.

For now apply hack so we can install beakerlib also on F30,
reported as bug rhbz#1719226.

Also, seems like systemd in container is broken again on F30;  but
we don't really need systemd (we actually wouldn't need even the
sleep because we don't need the container running all the time,
but I don't want to change the Makefile work-flow right now).