rpmbuild: make sure librepo/libdnf is always up2date
Librepo/libdnf is clearly part of the dnf/rpm family, so if we
expect that builders have the latest packaging stack -- we should
take care of those two (new) packages as well.

This commit is triggered by the need to update librepo/libdnf from
the bodhi update [1], which should solve major problems with
downloading mirrorlists, metalinks but individual RPMs as well.
For more info see [2, 3].

[1] https://bodhi.fedoraproject.org/updates/FEDORA-2019-13479b151a
[2] https://pagure.io/fedora-infrastructure/issue/7987
[3] https://github.com/rpm-software-management/librepo/pull/159