frontend: adds support for length of pending/running tables
In order to get length menu I had to update dom of DataTables.
It looks bit complicated, but when you look at documentation,
it should be clear.
Here is the link: https://datatables.net/reference/option/dom.

We decided to hide pagination using .js function when all table
records are shown. It is because of the DataTables bug where
pagination was showing wrong number of total pages. Istead of
"1 page out of 1 page", it was showing
"1 page out of (number_of_records) * -1".

Fixes #709