frontend, cli, python: support multilib projects
We create a new flask route for generating arch-specific repofiles.
Using this, and using the new per-project knob Project.multilib,
frontend can easily determine whether multilib repofile should be
generated or not.

For now, if the project is configured for multilib, we don't do any
magic on backend.  We simply provide repofile with two (32bit and 64bit)
repositories (the 32bit one contains `:ml` suffix in repoid).  And then
- if user does 'dnf copr enable' against multilib project - both
repositories are available on the system.

This trivial approach can be pretty easily enhanced in future,
especially if we wanted to support similar approach for per-package
multilib detection (as e.g. pungi has), we'd only have to drop the
second - multilib - variant of repo from repofile.

The fact that we are adding a new routenow, and that we have the `:ml`
suffix in repoid - we need to update the dnf-plugins-core package as
well [1].  But for compat with previous versions of dnf-plugins-core we
keep the old routes as well.

[1] https://github.com/rpm-software-management/dnf-plugins-core/pull/350

Fixes: #1
Merges: #938