frontend, backend: fixed project forking
This should fix broken project forking. In order to fix it, I had to
change the format of datas sending from frontend to backend.
Instead of lists containing names of old and new directories I am
sending dictionaries, where keys are old directories and values are
the names of new forked directories.

There is also unit test for this.

Fixes: #120, #652
Merges: #943