frontend, doc: more scripted fedora branching
This change adds `manage.py branch_fedora` command - that is a more
automated variant of manual rawhide_to_release command execution for
each rawhide chroot.  The change also introduces convenience option
`--deactivated` for `create_chroot` command, so we don't have to create
the chroot and disable right after that (racy).

Suggest in doc that we should newly use `branch_fedora` command, and
explain why is important to run the command as soon as possible.

Fixes: #955
Merges: #973