backend: do not provide the builder-live.log stub
Turns out that my original idea for PR#788 wasn't ideal and it is better
to _not_ provide builder-live.log stub.  Especially after
4aa9324f41adf6de9cd2 (UI generates appropriate link), so the http server
redirect configuration is mostly voluntary.

The thing is that it is very complicated (if possible at all) to
configure lighttpd (Fedora Copr deployment case) to redirect
automatically from 'builder-live.log' to 'builder-live.log.gz' when the
'builder-live.log.gz' _exists_.  And OTOH it is pretty trivial to
redirect to if 'builder-live.log' _doesn't exist_.

Also, notable benefit of this change is that we'll save some inode
quota.

While I'm on it, I'm adding single-shot script which will traverse the
build directory and (a) remove the stub logs we created before, and (b)
compress the non-stub uncompressed builder-live.log files (from the
times we did not compress at all).  This is needed to have working
'builder-live.log.gz' links from UI.

Fixes: #825
Relates: #921, PR#788, #941
Merges: #985