backend: avoid excception-hell in check_build_success()
Previously, we expected that Builder.check_build_success() raised
BuilderError in case of error, but this never happened - we instead
raised RemoteCmdError which was never catched.

Instead of cathing RemoteCmdError, execute `conn.run()` directly;  it is
both faster (we ignore stdout/stderr) and both we can avoid the
exception handling mesh in every call.

The only exception which could be raised there is SSHConnectionError,
which is carefully catched in worker.py.

Relates: #987